requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.1.0
paramiko==3.4.1
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
sqlparse==0.5.1
